Moved interfaces into a separate directory
