+++ Checkout.Tests/DiscountServiceTests.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using Checkout.Interfaces;
 using Moq;
 using NUnit.Framework;
