+++ Checkout/Checkout.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using Checkout.Exceptions;
+using Checkout.Interfaces;
 
 namespace Checkout
 {
