+++ Checkout/Interfaces/IBasket.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Checkout
+namespace Checkout.Interfaces
 {
     /// <summary>
     /// Represents a collection of <see cref="IProduct"/> ready for checkout
