+++ Checkout/Interfaces/IProduct.cs
@@ -1,4 +1,4 @@
-﻿namespace Checkout
+﻿namespace Checkout.Interfaces
 {
     /// <summary>
     /// Generic product interface
