+++ Checkout/Product.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Checkout.Interfaces;
 
 namespace Checkout
 {
